More gracefully handle end of series
Before seriesUID would be undefined when all cases have been
rated (or skipped) and that led to a never ending spinner and
lots of database timeout errors in the console.

Now the user is booted back to the login screen and if desired
can log in with a new account.