feat(leaderboard): add methods to get leaders by week
Adds:

* getTopAnnotatorsByWeek for individual scores for week,
defaults to current week

* getTopAnnotatorsByWeek for accumulated scores for a team per week,
defaults to current week

Both are included in getTopAnnotators but are not exported.

These should be integrated into the leaderboard template when
it is available (these functions can either be exported from
here or moved to dedicated source files to match style).

This uses the by/weekAnnotator view added to couchdb.